Fix and improve `createStream()` for creating a heartbeated stream
